requirements.txt
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-extensions==3.2.3
psycopg2==2.9.9
djoser==2.2.2
python-dotenv==1.0.1
youtube-dl==2021.12.17
yt-dlp==2023.12.30
django-unixdatetimefield==1.2.0
boto3==1.34.89
pillow==10.3.0